+++ lib/src/widgets/solid_login_build_helper.dart
@@ -46,6 +46,7 @@ class SolidLoginBuildHelper {
     required TextEditingController webIdController,
     required FocusNode focusNode,
   }) {
+    if (!style.visible) return const SizedBox.shrink();
     return FocusTraversalOrder(
       order: const NumericFocusOrder(3),
       child: SolidLoginButtons.buildRegisterButton(
@@ -69,6 +70,7 @@ class SolidLoginBuildHelper {
     required Future<void> Function() performLogin,
     required FocusNode focusNode,
   }) {
+    if (!style.visible) return const SizedBox.shrink();
     return FocusTraversalOrder(
       order: const NumericFocusOrder(1),
       child: SolidLoginButtons.buildLoginButton(
@@ -88,6 +90,7 @@ class SolidLoginBuildHelper {
     required Future<void> Function() performContinue,
     required FocusNode focusNode,
   }) {
+    if (!style.visible) return const SizedBox.shrink();
     return FocusTraversalOrder(
       order: const NumericFocusOrder(2),
       child: SolidLoginButtons.buildContinueButton(
@@ -105,6 +108,7 @@ class SolidLoginBuildHelper {
     required String link,
     required FocusNode focusNode,
   }) {
+    if (!style.visible) return const SizedBox.shrink();
     return FocusTraversalOrder(
       order: const NumericFocusOrder(4),
       child: SolidLoginButtons.buildInfoButton(