+++ lib/src/widgets/solid_scaffold_appbar_overflow.dart
@@ -56,6 +56,7 @@ class SolidAppBarOverflowHandler {
     SolidAboutConfig aboutConfig,
     BuildContext context, {
     bool showLogout = true,
+    bool showLogin = true,
     void Function(BuildContext)? onLogout,
     void Function(BuildContext)? onLogin,
   }) {